+++ HISTORY.md
@@ -4,3 +4,23 @@
 - NEW: Ability to switch between internal and external IP.
 
 * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+Version 1.0.0 (2015-04-24)
+========================================================================
+
+- CHANGE: Removed dependency on package 'python-yaml'.
+- CHANGED: Code was completely refactored.
+- FIX: Properly update menu when network state changes.
+- NEW: About dialog.
+- NEW: Automatically update IP whenever network connection changes.
+- NEW: IPs are displayed in menu as well.
+- NEW: Show all network interfaces in menu.
+
+* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * 
+
+
+Version 0.9.0 (2015-04-22)
+========================================================================
+
+- NEW: Ability to switch between internal and external IP.
+
+* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *